Refactor UserProfile
