+++ web/app/profile/[username]/page.tsx
@@ -3,7 +3,7 @@
 import ErrorMessage from "@/components/General/ErrorMessage";
 import LoadingSpinner from "@/components/LoadingSpinner";
 import PostCard from "@/components/PostCard";
-import UserProfile from "@/components/UserProfile";
+import UserProfile from "@/components/User/UserProfile";
 import WhoToFollow from "@/components/WhoToFollow";
 import WritePostCTA from "@/components/WritePostCTA";
 import useAuth from "@/hooks/useAuth";
